+++ packages/web3-core/src/web3_request_manager.ts
@@ -16,6 +16,7 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
 import { Socket } from 'net';
+
 import {
 	ContractExecutionError,
 	InvalidResponseError,
@@ -42,7 +43,7 @@ import {
 	Web3BaseProvider,
 	Web3BaseProviderConstructor,
 } from 'web3-types';
-import { isNullish, jsonRpc } from 'web3-utils';
+import { isNullish, isPromise, jsonRpc } from 'web3-utils';
 import {
 	isEIP1193Provider,
 	isLegacyRequestProvider,
@@ -200,34 +201,55 @@ export class Web3RequestManager<
 				);
 		}
 
-		// TODO: This should be deprecated and removed.
+		// TODO: This could be deprecated and removed.
 		if (isLegacyRequestProvider(provider)) {
-			return new Promise<JsonRpcResponse<ResponseType>>((resolve, reject): void => {
-				provider.request<ResponseType>(payload, (err, response) => {
-					if (err) {
-						return reject(
-							this._processJsonRpcResponse(
-								payload,
-								err as unknown as JsonRpcResponse<ResponseType>,
-								{
-									legacy: true,
-									error: true,
-								},
-							),
-						);
-					}
-
-					return resolve(
+			return new Promise<JsonRpcResponse<ResponseType>>((resolve, reject) => {
+				const rejectWithError = (err: unknown) =>
+					reject(
+						this._processJsonRpcResponse(
+							payload,
+							err as JsonRpcResponse<ResponseType>,
+							{
+								legacy: true,
+								error: true,
+							},
+						),
+					);
+				const resolveWithResponse = (response: JsonRpcResponse<ResponseType>) =>
+					resolve(
 						this._processJsonRpcResponse(payload, response, {
 							legacy: true,
 							error: false,
 						}),
 					);
-				});
+				const result = provider.request<ResponseType>(
+					payload,
+					// a callback that is expected to be called after getting the response:
+					(err, response) => {
+						if (err) {
+							return rejectWithError(err);
+						}
+
+						return resolveWithResponse(response);
+					},
+				);
+				// Some providers, that follow a previous drafted version of EIP1193, has a `request` function
+				//	that is not defined as `async`, but it returns a promise.
+				// Such providers would not be picked with if(isEIP1193Provider(provider)) above
+				//	because the `request` function was not defined with `async` and so the function definition is not `AsyncFunction`.
+				// Like this provider: https://github.dev/NomicFoundation/hardhat/blob/62bea2600785595ba36f2105564076cf5cdf0fd8/packages/hardhat-core/src/internal/core/providers/backwards-compatibility.ts#L19
+				// So check if the returned result is a Promise, and resolve with it accordingly.
+				// Note: in this case we expect the callback provided above to never be called.
+				if (isPromise(result)) {
+					const responsePromise = result as unknown as Promise<
+						JsonRpcResponse<ResponseType>
+					>;
+					responsePromise.then(resolveWithResponse).catch(rejectWithError);
+				}
 			});
 		}
 
-		// TODO: This should be deprecated and removed.
+		// TODO: This could be deprecated and removed.
 		if (isLegacySendProvider(provider)) {
 			return new Promise<JsonRpcResponse<ResponseType>>((resolve, reject): void => {
 				provider.send<ResponseType>(payload, (err, response) => {
@@ -261,7 +283,7 @@ export class Web3RequestManager<
 			});
 		}
 
-		// TODO: This should be deprecated and removed.
+		// TODO: This could be deprecated and removed.
 		if (isLegacySendAsyncProvider(provider)) {
 			return provider
 				.sendAsync<ResponseType>(payload)