+++ packages/web3/test/integration/external-providers/hardhat.test.ts
@@ -15,10 +15,14 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-// Have to use `require` because of Jest issue https://jestjs.io/docs/ecmascript-modules
-// eslint-disable-next-line @typescript-eslint/no-require-imports
-require('../config/setup');
+// eslint-disable-next-line import/no-extraneous-dependencies
+import hardhat from 'hardhat';
 
-const jestTimeout = 30000; // Sometimes `in3` takes long time because of its decentralized nature.
+import { performBasicRpcCalls } from './helper';
 
-jest.setTimeout(jestTimeout);
+describe('compatibility with `hardhat` provider', () => {
+	it('should initialize Web3, get accounts & block number and send a transaction', async () => {
+		// use the hardhat provider for web3.js
+		await performBasicRpcCalls(hardhat.network.provider);
+	});
+});